Fixed issues with variables not being accessed properly in toolbelt.js:
Ensured that uses of `doc` were initialized within their respective calling functions and set to `this.document`. Additionally, made it so that calls to `docId` were instead to `doc._id`.

Removed function calls that had void return being set to variables in reschedule/index.js

Made Operator a const so that it is more explicit as a variable for IDE parsing.